Move User#role_* methods through public RBAC interface
This privitizes the moved User logic in Authorizer and has any calls to
User go through RBAC's main interface instead.

It also adds some backwards compatible options we hadn't seen yet - for
example, some caller uses User#role_allows_any with 'identifiers',
something we hadn't been passing down previously. Also, 'scope' is an
option used by MiqUserRole internally, and was not being passed down.

This also changes the ControllerSpecHelper to stub
ApplicationController#role_allows instead of User. This may or may not
be the final answer in this stubbing bit but seems to help the tests
more than not, and also IS the way the controller tests should be
stubbing. So we'll see where that goes.

NOTE: BUILD IS RED AT THIS COMMIT