Allow UI spec suite to also compile assets.
Needs to verify the assets warnings are fixed or not if this is enabled
during test:

DEPRECATION WARNING: The asset "svg/vendor-physical_infra_manager.svg" is not present in the asset pipeline.Falling back to an asset that may be in the public folder.
This behavior is deprecated and will be removed.
To bypass the asset pipeline and preserve this behavior,
use the `skip_pipeline: true` option.
 (called from post_process_textual_summary at /home/travis/build/jrafanie/manageiq-ui-classic/app/helpers/textual_summary_helper.rb:40)