Add ActiveRecord::Type columns for not persisted objects to be serialized
Basically, active_record.use_yaml_unsafe_load = false requires these not yet
persisted objects be permitted.  We should not try to serialize unpersisted objects.

This test fails:
https://github.com/ManageIQ/manageiq-ui-classic/blob/7b6b73925027a05a5dac460b3a6c3df20d2e48e7/spec/controllers/application_controller/performance_spec.rb#L26-L39

Because, performance.rb tries to pass a MiqReport.new over the MiqQueue in args.  It's not persisted yet, so it's not using traditional classes
for columns, hence the need to add ActiveRecord::Type column types to the list of permitted classes.

https://github.com/ManageIQ/manageiq-ui-classic/blob/7b6b73925027a05a5dac460b3a6c3df20d2e48e7/app/controllers/application_controller/performance.rb#L570-L572