TERRIBLE HACK: Ignore calls to subclasses and descendants from two locations
Rails changed in 7.0 to call subclasses from reload_schema_from_cache here:
https://github.com/rails/rails/commit/6f30cc09a7ad21898d856059e498ab9ad2ae64a4

It also changed to call descendants on the callback class (self) insead of
the ActiveSupport::DescendantsTracker here:
https://github.com/rails/rails/commit/ffae3bd8d69f9ed1ae185e960d7a38ec17118a4d

We are not expecting to be called from these locations.  I don't know how to fix
this right now but want to leave a trail so future me or someone can have a stroke
of genius and completely fix this correctly.